requirements-dev.txt
amqp==5.1.1
asgiref==3.7.2
astroid==2.15.6
attrs==23.1.0
billiard==4.1.0
black==23.7.0
celery==5.3.1
cffi==1.15.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
dill==0.3.7
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
inflection==0.5.1
isort==5.12.0
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
kombu==5.3.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
pycparser==2.21
PyJWT==2.8.0
pylint==2.17.5
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.0
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.1
types-pyOpenSSL==23.2.0.2
types-redis==4.6.0.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
